pyproject.toml
[tool.poetry]
name = "hbrscrpr"
version = "1.0.0"
description = ""
authors = ["Alex Tyshkevich <alex@tyshkevich.ru>"]
readme = "README.md"

[tool.poetry.scripts]
hbrscrpr = "hbrscrpr.app:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.3.2"}
bs4 = "^0.0.1"
requests-mock = "^1.8.0"
tenacity = "^7.0.0"
responses = "^0.13.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
autoflake = "^1.4"
isort = "^5.7.0"
black = "^20.8b1"
unify = "^0.5"
flake8 = "^3.8.4"
pylint = "^2.7.1"
mypy = "^0.812"
pytest-deadfixtures = "^2.2.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
